refactor: Split id
